overhaul

+++ LIBS/LIBFT/ft_strlen.c
@@ -6,7 +6,7 @@
 /*   By: ccraciun <ccraciun@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/10/05 10:15:01 by corin             #+#    #+#             */
-/*   Updated: 2023/10/28 13:33:49 by ccraciun         ###   ########.fr       */
+/*   Updated: 2024/03/17 15:25:42 by ccraciun         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 size_t	ft_strlen(const char *str)
 {
-	int	len;
+	size_t	len;
 
 	len = 0;
 	while (*str)
